AOS-4599 rename AddGraphql to Add
